+++ src/main/java/com/programmers/epicblues/board/JpaPracticeBoardApplication.java
@@ -1,0 +1,15 @@
+package com.programmers.epicblues.board;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@SpringBootApplication
+@EnableJpaAuditing
+public class JpaPracticeBoardApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(JpaPracticeBoardApplication.class, args);
+  }
+
+}
