+++ src/main/java/com/programmers/epicblues/board/entity/BaseEntity.java
@@ -1,0 +1,20 @@
+package com.programmers.epicblues.board.entity;
+
+import java.time.LocalDateTime;
+import javax.persistence.Column;
+import javax.persistence.EntityListeners;
+import javax.persistence.MappedSuperclass;
+import lombok.Getter;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+@MappedSuperclass
+@Getter
+@EntityListeners(AuditingEntityListener.class)
+public abstract class BaseEntity {
+
+  protected String createdBy;
+  @CreatedDate
+  @Column(updatable = false)
+  protected LocalDateTime createdAt;
+}
