+++ src/main/java/com/programmers/epicblues/board/entity/LongIdHolder.java
@@ -1,0 +1,6 @@
+package com.programmers.epicblues.board.entity;
+
+public interface LongIdHolder {
+
+  Long getId();
+}
