+++ src/main/java/com/programmers/epicblues/board/exception/ResourceNotFoundException.java
@@ -1,0 +1,8 @@
+package com.programmers.epicblues.board.exception;
+
+public class ResourceNotFoundException extends RuntimeException {
+
+  public ResourceNotFoundException(String message, Throwable exception) {
+    super(message, exception);
+  }
+}
