+++ src/main/java/com/programmers/epicblues/board/repository/JpaUserRepository.java
@@ -1,0 +1,8 @@
+package com.programmers.epicblues.board.repository;
+
+import com.programmers.epicblues.board.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface JpaUserRepository extends JpaRepository<User, Long> {
+
+}
