--- src/main/java/com/programmers/epicblues/jpa_board/entity/LongIdHolder.java
@@ -1,6 +1,0 @@
-package com.programmers.epicblues.jpa_board.entity;
-
-public interface LongIdHolder {
-
-  Long getId();
-}
