--- src/main/java/com/programmers/epicblues/jpa_board/exception/InvalidRequestArgumentException.java
@@ -1,17 +1,0 @@
-package com.programmers.epicblues.jpa_board.exception;
-
-import org.springframework.validation.BindingResult;
-import org.springframework.validation.Errors;
-
-public class InvalidRequestArgumentException extends RuntimeException {
-
-  private final transient BindingResult bindingResult;
-
-  public InvalidRequestArgumentException(BindingResult bindingResult) {
-    this.bindingResult = bindingResult;
-  }
-
-  public Errors getBindingResult() {
-    return bindingResult;
-  }
-}
