--- src/main/java/com/programmers/epicblues/jpa_board/repository/JpaPostRepository.java
@@ -1,8 +1,0 @@
-package com.programmers.epicblues.jpa_board.repository;
-
-import com.programmers.epicblues.jpa_board.entity.Post;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface JpaPostRepository extends JpaRepository<Post, Long> {
-
-}
