--- src/main/java/com/programmers/epicblues/jpa_board/repository/JpaUserRepository.java
@@ -1,8 +1,0 @@
-package com.programmers.epicblues.jpa_board.repository;
-
-import com.programmers.epicblues.jpa_board.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface JpaUserRepository extends JpaRepository<User, Long> {
-
-}
